Move class-siwk-ajax.php to Ajax.php
